Fix another bug related to the channel banner
The bug was that when claiming a window, it is resized to its min
geometry automatically, which caused the channel banner pdf reader to be
resized to 480x360 first and then to 1920x1080 right after. This caused
an ugly page offset of the pdf reader when it showed the first or last
two pages of the pdf document that contains the channel banners.