Remove error message when broadcast has no receivers
When the plugin system starts, plugin instances broadcast module status
updates and messages updates. When neopult runs in an environment which
doesn't use stdin and thus doesn't use the terminal client (e.g. systemd
service), there are no receivers for that broadcast on startup. This
caused a false error message.