Add run_later lua api function
This is necessary for example, to run window manager api functions
inside a callback of the window manager.

Currently this limitation is imposed by the RwLock in the LuaContext
that holds the window manager instance. When an api function calls into
the window manager, the window manager has to be locked first. Now the
window manager could call callbacks on virtual windows (e.g. map/unmap).
When those callbacks call into the window manager again, it has to be
locked again. This leads to a panic because the current thread already
holds the lock.
Possible solution: Use interior mutability (e.g. RefCells) inside the
window manager and then remove the RwLock that wraps the window manager
in the lua context.