Improve design of `run_later` functionality
The new design processes all run_later tasks before new events are
processed. This prevents weird behavior that could happen otherwise. For
example, a plugin may call `run_later` with a function that sets a
window to min, while an user-triggered event is in the queue, that sets
that same window to max. First the user-triggered event would set the
window to max and shortly after, the window will be set to min by the
run_later task. This would appear to the user as though the action
was not applied correctly.