Add functionality to clean old processes
When spawning a process, the PID is saved into a channel-specific
directory in /tmp. The respective file is deleted when the process exits
while neopult is running. If the server should crash without being able
to kill its child processes, it can look up the PIDs on its next startup
and kill the old child processes.